install script with simplified conditionals

+++ install.sh
@@ -5,11 +5,7 @@ DIR="$(cd -P "$(dirname "$(readlink -f "${BASH_SOURCE[0]}")" )" && pwd)"
 
 # making a backup of the current screenrc the user might have
 TARGET="${HOME}/.screenrc"
-if [ -e "${TARGET}" ]; then  # only backup if it exists
-    if [ ! -L "${TARGET}" ]; then  # only if it is NOT a link
-        mv "${TARGET}" "${TARGET}.bkp"
-    fi
-fi
+[ -e "${TARGET}" ] && [ ! -L "${TARGET}" ] && mv "${TARGET}" "${TARGET}.bkp"
 
 # making the link to the version in this repo
 echo "source \"${DIR}/screenrc\"" > "${TARGET}"
